cap nhat

+++ Eatclean/src/main/java/eatclean/project/demo/repository/LoginRepository.java
@@ -11,7 +11,7 @@ import eatclean.project.demo.enity.User;
 
 
 @Repository
-public interface LoginRepository extends JpaRepository<Login, Long>{
+public interface LoginRepository extends JpaRepository<Login, Integer>{
 	Login findByUsernameAndPassword(String username, String password);
     Login findByUsername(String username);
     Login findById(int id);
